+++ libnodegl/node_render.c
@@ -149,7 +149,7 @@ static const struct node_param render_params[] = {
     {"instance_attributes", NGLI_PARAM_TYPE_NODEDICT, OFFSET(instance_attributes),
                  .node_types=ATTRIBUTES_TYPES_LIST,
                  .desc=NGLI_DOCSTRING("per instance extra vertex attributes made accessible to the `program`")},
-    {"nb_instances", NGLI_PARAM_TYPE_I32, OFFSET(nb_instances), {.i64 = 1},
+    {"nb_instances", NGLI_PARAM_TYPE_I32, OFFSET(nb_instances), {.i32 = 1},
                  .desc=NGLI_DOCSTRING("number of instances to draw")},
     {"blending", NGLI_PARAM_TYPE_SELECT, OFFSET(blending),
                  .choices=&ngli_blending_choices,
